refactoring plugin_sge.py and others
