+++ plugin_oar.py
@@ -49,9 +49,9 @@ class OarStatMaker(QStatMaker):
                              r'(?P<job_state>[RWF])\s+' \
                              r'(?P<queue>default|besteffort)'
 
-    def convert_qstat_to_yaml(self, orig_file, out_file, write_method):
+    def serialise_qstat(self, orig_file, out_file, write_method):
         with open(orig_file, 'r') as fin:
-            logging.debug('File state before OarStatMaker.convert_qstat_to_yaml: %(fin)s' % {"fin": fin})
+            logging.debug('File state before OarStatMaker.serialise_qstat: %(fin)s' % {"fin": fin})
             _ = fin.readline()  # header
             fin.readline()  # dashes
             re_match_positions = ('job_id', 'user', 'job_state', 'queue')
@@ -60,7 +60,7 @@ class OarStatMaker(QStatMaker):
                 qstat_values = self.process_line(re_search, line, re_match_positions)
                 self.l.append(qstat_values)
 
-        logging.debug('File state after OarStatMaker.convert_qstat_to_yaml: %(fin)s' % {"fin": fin})
+        logging.debug('File state after OarStatMaker.serialise_qstat: %(fin)s' % {"fin": fin})
         self.dump_all(out_file, self.stat_mapping[write_method])
